fix: sort record batches when persist to parquet
