+++ src/config/Cargo.toml
@@ -24,6 +24,7 @@ byteorder.workspace = true
 chrono.workspace = true
 cityhasher.workspace = true
 chromiumoxide.workspace = true
+datafusion.workspace = true
 dashmap.workspace = true
 dotenv_config.workspace = true
 dotenvy.workspace = true
