+++ src/service/compact/merge.rs
@@ -29,7 +29,7 @@ use config::{
         parquet::{
             parse_file_key_columns, read_recordbatch_from_bytes, write_recordbatch_to_parquet,
         },
-        record_batch_ext::{concat_batches, format_recordbatch_by_schema},
+        record_batch_ext::{format_recordbatch_by_schema, merge_record_batchs},
     },
     FILE_EXT_PARQUET,
 };
@@ -1069,11 +1069,11 @@ pub fn generate_inverted_idx_recordbatch(
 pub async fn merge_parquet_files(
     thread_id: usize,
     trace_id: &str,
-    mut schema: Arc<Schema>,
+    schema: Arc<Schema>,
 ) -> ::datafusion::error::Result<(Arc<Schema>, Vec<RecordBatch>)> {
     let start = std::time::Instant::now();
 
-    // 1. get record batches from tmpfs
+    // get record batches from tmpfs
     let temp_files = infra::cache::tmpfs::list(trace_id, "parquet").map_err(|e| {
         log::error!(
             "[MERGE:JOB:{thread_id}] merge small files failed at getting temp files. Error {}",
@@ -1110,64 +1110,13 @@ pub async fn merge_parquet_files(
         .map(|b| format_recordbatch_by_schema(schema.clone(), b))
         .collect::<Vec<_>>();
 
-    // 2. concatenate all record batches into one single RecordBatch
-    let mut concated_record_batch = concat_batches(
-        schema.clone(),
-        record_batches,
-        get_config().compact.fast_mode,
-    )?;
-
-    // 3. delete all the null columns
-    let num_rows = concated_record_batch.num_rows();
-    let mut null_columns = Vec::new();
-    for idx in 0..schema.fields().len() {
-        if concated_record_batch.column(idx).null_count() == num_rows {
-            null_columns.push(idx);
-        }
-    }
-    if !null_columns.is_empty() {
-        for (deleted_num, idx) in null_columns.into_iter().enumerate() {
-            concated_record_batch.remove_column(idx - deleted_num);
-        }
-        schema = concated_record_batch.schema().clone();
-    }
-
-    // 4. sort concatenated record batch by timestamp col in desc order
-    let sort_indices = arrow::compute::sort_to_indices(
-        concated_record_batch
-            .column_by_name(&get_config().common.column_timestamp)
-            .ok_or_else(|| {
-                log::error!(
-                    "[MERGE:JOB:{thread_id}] merge small files failed to find _timestamp column from merged record batch.",
-                );
-                DataFusionError::Execution(
-                    "No _timestamp column found in merged record batch".to_string(),
-                )
-            })?,
-        Some(arrow_schema::SortOptions {
-            descending: true,
-            nulls_first: false,
-        }),
-        None,
-    )?;
-
-    let batch_columns_len = concated_record_batch.columns().len();
-    let mut sorted_columns = Vec::with_capacity(batch_columns_len);
-    for i in 0..batch_columns_len {
-        let i = i - sorted_columns.len();
-        let sorted_column =
-            arrow::compute::take(&concated_record_batch.remove_column(i), &sort_indices, None)?;
-        sorted_columns.push(sorted_column);
-    }
-    drop(concated_record_batch);
-
-    let final_record_batch = RecordBatch::try_new(schema.clone(), sorted_columns)?;
-    let schema = final_record_batch.schema();
+    // merge record batches, the record batch have same schema
+    let (schema, record_batches) = merge_record_batchs("MERGE", thread_id, schema, record_batches)?;
 
     log::info!(
         "[MERGE:JOB:{thread_id}] merge_parquet_files took {} ms",
         start.elapsed().as_millis()
     );
 
-    Ok((schema, vec![final_record_batch]))
+    Ok((schema, record_batches))
 }
